cpufreq: Adaptive CPU governor
This driver adds a dynamic cpufreq policy governor designed for
latency-sensitive workloads and also for demanding performance.
This governor attempts to reduce the latency of clock increases so
that the system is more responsive to interactive workloads in loweset
steady-state but to reduce power consumption in middle operation level
up will be done in step by step to prohibit system from going to max
operation level.